+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Implemented parameter groups!
+- Added `EdgesForceDensityParameter`.
+- Added `NodesLoadXParameter`, `NodesLoadYParameter`, `NodesLoadZParameter`.
+- Added `NodesSupportXParameter`, `NodesSupportYParameter`, `NodesSupportZParameter`.
 - Implemented `area_polygon` in `jax_fdm.geometry`.
 - Added `FDNetwork.number_of_supports()` to count number of supported nodes in a network.
 - Added `network_validate` to check the validity of a network before equilibrium calculations.
@@ -23,7 +27,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - Removed `LossPlotter.network` attribute.
-
+- Replaced `NodeAnchorXParameter` with `NodeSupportXParameter`.
+- Replaced `NodeAnchorYParameter` with `NodeSupportYParameter`.
+- Replaced `NodeAnchorZParameter` with `NodeSupportZParameter`.
 
 ## [0.4.5] 2023-02-04